Refactoring the extraction code to make it more digestible and testable
