+++ hles/extraction/src/main/scala/org/broadinstitute/monster/dap/ExtractionPipelineBuilder.scala
@@ -38,15 +38,15 @@ object ExtractionPipelineBuilder {
   *                  interact with a RedCap API
   */
 class ExtractionPipelineBuilder(
-  formsForExtraction: List[String],
-  extractionFilters: List[FilterDirective],
-  arm: List[String],
-  fieldList: List[String],
-  subDir: String,
-  idBatchSize: Int,
-  getClient: List[String] => RedCapClient
-) extends PipelineBuilder[Args]
-    with Serializable {
+                                 formsForExtraction: List[String],
+                                 extractionFilters: List[FilterDirective],
+                                 arm: List[String],
+                                 fieldList: List[String],
+                                 subDir: String,
+                                 idBatchSize: Int,
+                                 getClient: (List[String], HttpWrapper) => RedCapClient
+                               ) extends PipelineBuilder[Args]
+  with Serializable {
 
   override def buildPipeline(ctx: ScioContext, args: Args): Unit = {
     import org.broadinstitute.monster.common.msg.MsgOps
@@ -56,10 +56,11 @@ class ExtractionPipelineBuilder(
       new ScalaAsyncLookupDoFn[RedcapRequest, Msg, RedCapClient](MaxConcurrentRequests) {
         override def newClient(): RedCapClient = getClient(arm)
         override def asyncLookup(
-          client: RedCapClient,
-          input: RedcapRequest
-        ): Future[Msg] =
-          client.get(args.apiToken, input)
+                                  client: RedCapClient,
+                                  input: RedcapRequest
+                                ): Future[Msg] =
+          // todo pass in a client here?
+          client.get(args.apiToken, input, httpClient)
       }
 
     // Start by pulling the IDs that match the supplied filtering criteria
@@ -70,8 +71,12 @@ class ExtractionPipelineBuilder(
       filters = extractionFilters
     )
     val idsToExtract = ctx
+      // massaging data to get back an SCollection[]
       .parallelize(Iterable(initRequest))
       .transform("Get study IDs") {
+        // Passing the lookup function (RedCap Future[Msg]) and applying the transform across
+        // Defining a transformation for a key value pair (study_id, value) (Steve's understanding)
+        // get back a list of obj each of which has a key called "value" (FLATTENING the object)
         _.applyKvTransform(ParDo.of(lookupFn)).flatMap { kv =>
           kv.getValue.fold(throw _, _.arr.map(_.read[String]("value")))
         }
@@ -81,6 +86,8 @@ class ExtractionPipelineBuilder(
     // NOTE: This logic is replicated from the encode-ingest pipeline,
     // we should consider moving it to scio-utils.
     val batchedIds = idsToExtract
+      // static initializer for this kv
+      // message serialization into uPack Msg
       .map(KV.of("", _))
       .setCoder(KvCoder.of(StringUtf8Coder.of(), StringUtf8Coder.of()))
       .applyKvTransform(GroupIntoBatches.ofSize(idBatchSize.toLong))
